feat : feature create view

+++ database/migrations/2022_04_18_133023_create_match_strings_table.php
@@ -18,7 +18,7 @@ class CreateMatchStringsTable extends Migration
             $table->string('match');
             $table->text('text');
             $table->integer('matching');
-            $table->decimal('total');
+            $table->integer('total');
             $table->timestamps();
         });
     }
